% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_ps_gmm_dg}
\alias{c_ps_gmm_dg}
\title{Get Derivative of Moment Constraints}
\usage{
c_ps_gmm_dg(beta, mat_grp_x, att = FALSE)
}
\description{
return k row p column matrix
Row:    1..k moment functions
Column: 1..p coefficients
}
