% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_environment.R
\name{clean_environment}
\alias{clean_environment}
\title{Clean the R Environment}
\usage{
clean_environment(env = .GlobalEnv, exclude = character())
}
\arguments{
\item{env}{An environment from which to remove data frames. Default
`.GlobalEnv`.}

\item{exclude}{A character vector of data frame names to keep. Default
`character()`.}
}
\value{
Invisibly returns `NULL`. Called for its side effect of removing
  objects from `env`.
}
\description{
Removes all data frames from the specified environment, optionally
excluding named objects.
}
\examples{
\dontrun{
df1 <- data.frame(x = 1)
df2 <- data.frame(y = 2)
df3 <- data.frame(z = 3)

# Remove all data frames
clean_environment()

# Keep df1
clean_environment(exclude = "df1")
}
}
