% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cMatch}
\alias{cMatch}
\title{Match by nearest neighbor}
\usage{
cMatch(target, candidate, ratio)
}
\arguments{
\item{target}{vector of data from the target group}

\item{ratio}{1:ratio match}

\item{candidates}{vector of data from the candidate group}
}
\description{
Match subjects in group candidate with subject in the target group
}
