% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_package.R
\name{ex_data}
\alias{ex_data}
\title{Example dataset}
\format{
A dataframe with the following variables:
\describe{
  \item{group}{0,1,2}
  \item{V1-V10}{Covariates}
  \item{Y}{Binary outcome}
}
}
\description{
This is a simulated dataset
}
