% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\docType{data}
\name{gem_10_to_9}
\alias{gem_10_to_9}
\title{General equivalence mapping from ICD10CM to ICD9CM (2018 edition)}
\format{
A data frame
\describe{
  \item{icd10}{ICD10CM code.}
  \item{icd9}{ICD9CM code.}
  ...
}
}
\usage{
gem_10_to_9
}
\description{
A dataset mapping ICD10CM codes to ICD9CM codes (2018 edition)
}
\keyword{datasets}
