% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\docType{data}
\name{gem_9_to_10}
\alias{gem_9_to_10}
\title{General equivalence mapping from ICD9CM to ICD10CM (2018 edition)}
\format{
A data frame
\describe{
  \item{icd9}{ICD9CM code.}
  \item{icd10}{ICD10CM code.}
  ...
}
}
\usage{
gem_9_to_10
}
\description{
A dataset mapping ICD9CM codes to ICD10CM codes (2018 edition)
}
\keyword{datasets}
