% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_balance.R
\name{get_metric}
\alias{get_metric}
\title{Get balance by different metrics}
\usage{
get_metric(cov0, cov1, metric = c("std", "abd", "ovl", "ksd", "ley", "mhb"))
}
\arguments{
\item{metric}{different metrics for measuring balance}
}
\description{
Get balance by different metrics
}
