% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\docType{data}
\name{icd10_2019}
\alias{icd10_2019}
\title{ICD10CM medical codes (2019 edition)}
\format{
A data frame
\describe{
  \item{code}{ICD10CM code.}
  \item{leaf}{Logical indicating whether code is billable}
  \item{short_desc}{Short literal description of code}
  ...
}
}
\usage{
icd10_2019
}
\description{
A dataset of ICD10CM codes for 2019. Courtesy of the `icd` package.
}
\keyword{datasets}
