% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data.R
\docType{data}
\name{icd9_2014}
\alias{icd9_2014}
\title{ICD9CM medical codes (2014 edition)}
\format{
A data frame
\describe{
  \item{code}{ICD9CM code.}
  \item{leaf}{Logical indicating whether code is billable}
  \item{short_desc}{Short literal description of code}
  ...
}
}
\usage{
icd9_2014
}
\description{
A dataset of ICD9CM codes for 2014. Courtesy of the `icd` package.
}
\keyword{datasets}
