% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_balance.R
\name{metric_mhb}
\alias{metric_mhb}
\title{mahalanobis balance}
\usage{
metric_mhb(cov0, cov1)
}
\description{
covs should be a reduced datset that contains only those covariates
that will be used for calculating Mahalanobis balance, for example,
covs = dat[,1:6]
trt should be the exposure variable,
for example, trt=dat$X
}
