% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hc_treemap.R
\name{plot_hc_treemap}
\alias{plot_hc_treemap}
\title{Interactive Treemap (Highcharts)}
\usage{
plot_hc_treemap(data, group_by, decimals = 2)
}
\arguments{
\item{data}{A data frame containing the data to be summarised and plotted.}

\item{group_by}{A string specifying the column to group by (defines tile
labels).}

\item{decimals}{Integer number of decimal places shown in the tooltip
percentage. Default `2`.}
}
\value{
A `highchart` object.
}
\description{
Generates an interactive treemap chart using `highcharter`, where tile size
and colour both reflect the count of records per group.
}
\examples{
\dontrun{
data("mpg", package = "ggplot2")
plot_hc_treemap(data = mpg, group_by = "manufacturer")
}
}
