% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_margin.R
\name{private_margin_ll}
\alias{private_margin_ll}
\title{Calculate log likelihood}
\usage{
private_margin_ll(lst_stats, y)
}
\arguments{
\item{lst_stats}{summary statistics}

\item{y}{vector of data}
}
\value{
column vector with n patients
}
\description{
Calculate likelihood of a covariate based on its summary statistics
}
