% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_margin.R
\name{private_margin_simu}
\alias{private_margin_simu}
\title{Simulate a covariate based on summary statistics}
\usage{
private_margin_simu(lst_stats, n = 500)
}
\value{
column vector with n patients
}
\description{
Simulate a covariate based on summary statistics
}
