% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_margin.R
\name{private_quantile_mat}
\alias{private_quantile_mat}
\title{Quantile matrix}
\usage{
private_quantile_mat(range, quants)
}
\value{
matrix with 3 columns. 1: probabilities, 2: lower bound,
3: upper bound
}
\description{
Get matrix data of quantiles
}
