% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_balance.R
\name{rwe_bal_metric}
\alias{rwe_bal_metric}
\title{Get balance by different metrics}
\usage{
rwe_bal_metric(covs, grp, metric = c("std", "abd", "ovl", "ksd", "ley", "mhb"))
}
\arguments{
\item{covs}{covariate matrix}

\item{grp}{group indicator}

\item{metric}{different metrics for measuring balance}
}
\description{
Get balance by different metrics
}
