% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_balance.R
\name{rwe_bal_metric_ps}
\alias{rwe_bal_metric_ps}
\title{Get PS-adjusted balance by different metrics}
\usage{
rwe_bal_metric_ps(
  covs,
  grp,
  ps,
  adjust = c("matching", "strat"),
  n_strata = 5,
  m_ratio = 3,
  ...
)
}
\arguments{
\item{covs}{covariate matrix}

\item{grp}{group indicator}
}
\description{
Get PS-adjusted balance by different metrics
}
