% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_margin.R
\name{rwe_extract_stats}
\alias{rwe_extract_stats}
\title{Extract summary statistics}
\usage{
rwe_extract_stats(target_stats, y)
}
\value{
list of summary statistics
}
\description{
Extract summary statistics from a data frame based on an existing summary
statistics list
}
