% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_margin.R
\name{rwe_extract_stats_covx}
\alias{rwe_extract_stats_covx}
\title{Get estimating equation}
\usage{
rwe_extract_stats_covx(target_stats, y, cont_stat = c("mean", "ex2"))
}
\value{
list of summary statistics
}
\description{
Get moment equations based on existing statistics
}
