% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_ps.R
\name{rwe_gmm_ps_old}
\alias{rwe_gmm_ps_old}
\title{Get Propensity Scores by GMM}
\usage{
rwe_gmm_ps_old(
  grp,
  ps_fml,
  dta,
  att = 0,
  d1_grp = 1,
  method = "BFGS",
  crit = 1e-10,
  tol = 1e-08,
  itermax = 5000,
  control = list(reltol = 1e-10, maxit = 20000),
  ...
)
}
\description{
Estimating PS by GMM for three groups
}
