% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_margin.R
\name{rwe_margin_entropy}
\alias{rwe_margin_entropy}
\title{Calculate entropy score}
\usage{
rwe_margin_entropy(
  target_stats,
  dta_ext,
  tol = 1e-08,
  max_it = 10000,
  print_level = 0
)
}
\description{
Get entropy balancing score
}
