% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_margin.R
\name{rwe_margin_ll}
\alias{rwe_margin_ll}
\title{Calculate log-likelihood based on summary statistics}
\usage{
rwe_margin_ll(target_stats, y)
}
\arguments{
\item{y}{matrix of data, each column represents a covariate and each row
represents a patient}
}
\value{
data frame for all y
}
\description{
Calculate log-likelihood based on summary statistics
}
