% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_margin.R
\name{rwe_margin_ps}
\alias{rwe_margin_ps}
\title{Calculate PS score based on simulated target data}
\usage{
rwe_margin_ps(target_stats, dta_ext, n_cur = NULL, reps = 10, ps_thresh = 0.01)
}
\description{
Calculate PS score based on simulated target data
}
