% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_margin.R
\name{rwe_margin_simu}
\alias{rwe_margin_simu}
\title{Simulate covariates based on summary statistics}
\usage{
rwe_margin_simu(target_stats, n = 500)
}
\value{
data frame with n patients, each column represents a covariate and
    each row represents a patient
}
\description{
Simulate covariates based on summary statistics
}
