% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_margin.R
\name{rwe_margin_stat_diff}
\alias{rwe_margin_stat_diff}
\title{Calculate differences in summary statistics}
\usage{
rwe_margin_stat_diff(lst_stats, y, type = c("max", "sum"),
  epsilon = NULL, ...)
}
\value{
vector of differences in summary statistics
}
\description{
Calculate differences in summary statistics
}
