% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_ps.R
\name{rwe_match_ps}
\alias{rwe_match_ps}
\title{Match on PS}
\usage{
rwe_match_ps(dta_cur, dta_ext, n_match = 3, ps.fml = NULL, v.covs = "V1")
}
\arguments{
\item{dta_cur}{current study data}

\item{dta_ext}{external data source data}
}
\description{
Match patients in RWD with patients in current study based on PS
}
