% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_surv.R
\name{rwe_ps_surv}
\alias{rwe_ps_surv}
\title{PS-Integrated Survival Curve}
\usage{
rwe_ps_surv(
  data,
  lambdas,
  v_censored = "censored",
  v_event = "time",
  m_var = c("jk"),
  pred_tps = 1,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{class DWITHPS data frame}

\item{lambdas}{power parameter without standardization by ns0}

\item{m_var}{method to get variance: jackknife or bootstrap}

\item{seed}{random seed}

\item{...}{parameters for \code{rweWL}}
}
\description{
For all stratum. Variance estimated by Jack-Knife.
}
