% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rwe_simu_surv}
\alias{rwe_simu_surv}
\title{Simulate survival outcomes}
\usage{
rwe_simu_surv(
  nPat,
  muCov,
  sdCov,
  corCov,
  b0,
  regCoeff,
  muCov_cens,
  sdCov_cens,
  corCov_cens,
  b0_cens,
  regCoeff_cens,
  fmla_surv = NULL,
  fmla_cens = NULL,
  t_max = 5,
  ...
)
}
\description{
Simulate survival outcomes based on exponential distribution
}
