% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_ps.R
\name{rwe_tk_ps}
\alias{rwe_tk_ps}
\title{Get Propensity Scores}
\usage{
rwe_tk_ps(
  dta,
  ps_fml = NULL,
  d1_grp = 1,
  type = c("logistic", "randomforest", "gmm"),
  ...,
  grp = NULL,
  ps_cov = NULL,
  ntree = 5000
)
}
\description{
General function for estimating PS
}
