% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_balance.R
\name{rweBalMetric}
\alias{rweBalMetric}
\title{Get balance by different metrics}
\usage{
rweBalMetric(cov0, cov1, metric = c("std", "abd", "ovl", "ksd", "ley",
  "mhb"))
}
\arguments{
\item{cov0}{covariates from group 0}

\item{cov1}{covariates from group 1}
}
\description{
Get balance by different metrics
}
