% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_tools.R
\name{rweEvenLmbdS}
\alias{rweEvenLmbdS}
\title{Split A into Bins}
\usage{
rweEvenLmbdS(ns1.trt, ns1.ctl, ns0, A, init.lmbds = NULL)
}
\arguments{
\item{ns1.trt}{numbers of treatment arm patients in each stratum}

\item{ns1.ctl}{numbers of control arm patients in each stratum}

\item{ns0}{numbers of external patients in each stratum}

\item{A}{number of target patients to borrow}
}
\description{
Split A into bins to minimize the number difference between two arms
}
