% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_present.R
\name{rweExpRst}
\alias{rweExpRst}
\title{Export results into a template file}
\usage{
rweExpRst(numbers, template.f, out.f = "rst.txt", sub.str = "AA")
}
\arguments{
\item{numbers}{vector of results}

\item{template.f}{template file name}

\item{out.f}{output file name}

\item{sub.str}{pattern of string to be replaced}
}
\description{
Export results into a template file
}
