% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweGetBinInt}
\alias{rweGetBinInt}
\title{Get intercept for a binary outcome.}
\usage{
rweGetBinInt(bin.mu, nPat = 10000, xbeta = NULL, ...)
}
\arguments{
\item{bin.mu}{mean of the binary outcomes used to compute b0}

\item{nPat}{number of patients}
}
\value{
standard error or the random error term
}
\description{
The binary outcome may be an outcome or a treatment assignment.
}
