% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweGetYSig}
\alias{rweGetYSig}
\title{Compute standard error of the random error}
\usage{
rweGetYSig(sig2Ratio, nPat = 10000, xbeta = NULL, ...)
}
\arguments{
\item{nPat}{number of patients}
}
\value{
standard error or the random error term
}
\description{
Compute standard error of the random error
}
