% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_ps.R
\name{rweGPS}
\alias{rweGPS}
\title{Get generalized propensity scores}
\usage{
rweGPS(data, ps.fml = NULL, v.grp = "group", v.covs = "V1",
  nstrata = 5, ...)
}
\arguments{
\item{...}{parameters to get propensity scores}
}
\description{
Get generalized propensity scores
}
