% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_tools.R
\name{rweGpsLambda}
\alias{rweGpsLambda}
\title{Get weights}
\usage{
rweGpsLambda(A, ps_dist)
}
\arguments{
\item{A}{target number of subjects to be borrowed}

\item{m.lambda}{method to split A. rs: by overlapping coefficient; even: by
minimizing trt and control imbalance in numbers}
}
\value{
power parameter before standardization
}
\description{
Get weights
}
