% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_tools.R
\name{rweKL}
\alias{rweKL}
\title{Compute Kullback-Leibler distance from F0 to F1}
\usage{
rweKL(sample.F0, sample.F1, n.bins = 10, epsilon = 10^-6)
}
\arguments{
\item{n.bins}{number of bins for KL computation}

\item{epsilon}{small integer for Dirichlet smoothing}
}
\value{
a vector with the number of samples in group 0, the number of samples
    in group 1, and the KL divergence from group 0 to group 1
}
\description{
Compute Kullback-Leibler distance from F0 to F1
}
