% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_present.R
\name{rwePlotUnbalance}
\alias{rwePlotUnbalance}
\title{Plot unbalance in covariates}
\usage{
rwePlotUnbalance(data.unb, xlim = NULL, ylim = NULL, title = "",
  f.grid = formula("V~Study"))
}
\description{
Plot unbalance in covariates
}
