% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_wl.R
\name{rwePsJkWL}
\alias{rwePsJkWL}
\title{PS-Integrated Composite Likelihood Estimation for all
Jack-Knife values}
\usage{
rwePsJkWL(data, lambdas, v.outcome = "Y", ...)
}
\arguments{
\item{data}{class DWITHPS data frame}

\item{lambdas}{power parameter without standardization by ns0}

\item{...}{parameters for \code{rweWL}}

\item{seed}{random seed}
}
\description{
PS-Integrated Composite Likelihood Estimation for all
Jack-Knife values
}
