% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_ps.R
\name{rwePSKL}
\alias{rwePSKL}
\title{Get number of subjects and the KL distance for each PS strata}
\usage{
rwePSKL(data.withps, ...)
}
\description{
Get number of subjects and the KL distance for each PS strata
}
