% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_stan.R
\name{rwePsPowDrawPost}
\alias{rwePsPowDrawPost}
\title{Get Posterior for all stratum}
\usage{
rwePsPowDrawPost(data.ps, type = c("continuous", "binary"), A = 0,
  RS = NULL, v.outcome = "Y", ...)
}
\arguments{
\item{data.ps}{class DWITHPS data frame}

\item{type}{type of outcomes}

\item{A}{target number of subjects to be borrowed}

\item{RS}{parameters for dirichelet prior}

\item{...}{extra parameters for calling function \code{\link{rweSTAN}}}
}
\description{
Get Posterior for all stratum
}
