% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_stan.R
\name{rwePsPowDrawPrior}
\alias{rwePsPowDrawPrior}
\title{Get Prior Only for all stratum}
\usage{
rwePsPowDrawPrior(data, type = c("continuous", "binary"), A = 0,
  RS = NULL, v.outcome = "Y", iter = 2000, ...)
}
\arguments{
\item{data}{class DWITHPS data frame}

\item{type}{type of outcomes}

\item{A}{target number of subjects to be borrowed}

\item{RS}{parameters for dirichelet prior}

\item{...}{extra parameters for calling function \code{\link{rweSTAN}}}
}
\description{
Get Prior Only for all stratum
}
