% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_balance.R
\name{rwePSUnbalance}
\alias{rwePSUnbalance}
\title{Get unbalance after PS adjustment}
\usage{
rwePSUnbalance(
  pts,
  covs = NULL,
  diff = TRUE,
  var.group = "Z",
  cov.pattern = "^V[0-9]+$"
)
}
\arguments{
\item{pts}{data frame of datasets}

\item{var_group}{Column name in pts that corresponds to propensity score}
}
\value{

}
\description{
Get unbalance after PS adjustment
}
