% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_wl.R
\name{rwePsWL}
\alias{rwePsWL}
\title{PS-Integrated Weighted Likelihood Estimation for all stratum}
\usage{
rwePsWL(data, RS = NULL, A = 0, v.outcome = "Y", bs.n = 1000,
  bs.conf = c(0.025, 0.975), seed = NULL, ...)
}
\arguments{
\item{data}{class DWITHPS data frame}

\item{RS}{parameters for dirichelet prior}

\item{A}{target number of subjects to be borrowed}

\item{bs.n}{number of bootstraps}

\item{bs.conf}{bootstrap confidence interval level}

\item{seed}{random seed}

\item{...}{parameters for \code{rweWL}}
}
\description{
PS-Integrated Weighted Likelihood Estimation for all stratum
}
