% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_shiny.R
\name{rweShiny}
\alias{rweShiny}
\title{Run Web-Based \code{rwetools} application}
\usage{
rweShiny()
}
\description{
Call Shiny to run \code{rwetools} as a web-based application. A web browser will
be brought up.
}
\examples{
\dontrun{
rweShiny()}


}
