% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweSimuCov}
\alias{rweSimuCov}
\title{Simulate covariates}
\usage{
rweSimuCov(nPat, muCov, sdCov, corCov)
}
\arguments{
\item{nPat}{number of patients}

\item{muCov}{mean vector of covariates}

\item{sdCov}{standard deviation vector of covariates}

\item{corCov}{correlation of covariates}
}
\description{
Simulate covariates
}
