% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweSimuFromTrial}
\alias{rweSimuFromTrial}
\title{Simulate data from an existing dataset}
\usage{
rweSimuFromTrial(nPat, trial.data, group = "A", outcome = "Y",
  keep.group = TRUE, trt.effect = 0, with.replacement = FALSE,
  ratio.size = 1, seed = NULL)
}
\arguments{
\item{nPat}{number of patients}

\item{trial.data}{existing clinical trial data}

\item{group}{column referring to arm in the existing dataset}

\item{outcome}{column referring to outcome in the existing dataset}

\item{keep.group}{sample ignore arm assignment in the existing dataset}

\item{with.replacement}{sample with or without replacement from the existing dataset}
}
\description{
Simulate data from an existing dataset
}
