% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweSimuOneArm}
\alias{rweSimuOneArm}
\title{Simulate outcomes for a single arm study}
\usage{
rweSimuOneArm(nPat, muCov, sdCov, corCov, regCoeff, fmla = NULL,
  type = c("continuous", "binary"), ysig = NULL, sig2Ratio = 2,
  bin.mu = 0.5, ...)
}
\arguments{
\item{nPat}{number of patients}

\item{muCov}{mean vector of covariates}

\item{sdCov}{standard deviation vector of covariates}

\item{corCov}{correlation of covariates}

\item{regCoeff}{regression coefficients}

\item{type}{distributions of the random error}

\item{ysig}{standard error of the random error}

\item{bin.mu}{mean of the binary outcomes used to compute b0}
}
\description{
Simulate outcomes for a single arm study
}
