% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_package.R
\docType{package}
\name{rwetools-package}
\alias{rwetools-package}
\alias{rwetools}
\title{R tools for synthesizing real-world evidence}
\description{
This package contains the functions for synthesizing real-world evidence in
single-arm medical device studies.
}
