% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweUnbalance}
\alias{rweUnbalance}
\title{Get unbalance in baseline covariates}
\usage{
rweUnbalance(nPat, ..., pts = NULL)
}
\arguments{
\item{nPat}{number of patients}
}
\description{
Get unbalance in baseline covariates
}
