% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_wl.R
\name{rweWL}
\alias{rweWL}
\title{Weighted Likelihood Estimation for one external RWD}
\usage{
rweWL(cur.data, ext.data, lambda, type = c("continuous", "binary"),
  equal.sd = TRUE)
}
\arguments{
\item{cur.data}{data from current study}

\item{ext.data}{data from external study}

\item{lambda}{power parameter without standardization of ns0}

\item{type}{type of outcomes}

\item{equal.sd}{boolean. whether sd is the same between the current and external study}
}
\description{
Weighted Likelihood Estimation for one external RWD
}
