% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_wl.R
\name{rweWL2}
\alias{rweWL2}
\title{Weighted Likelihood Estimation}
\usage{
rweWL2(cur.data, ext.data, ext.ps, type = c("continuous", "binary"),
  equal.sd = FALSE)
}
\arguments{
\item{cur.data}{data from current study}

\item{ext.data}{data from external study}

\item{ext.ps}{ps of the external study}

\item{type}{type of outcomes}

\item{equal.sd}{boolean. whether sd is the same between the current and external study}
}
\description{
Weighted Likelihood Estimation
}
