% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_wl.R
\name{rweWL_G}
\alias{rweWL_G}
\title{Weighted Likelihood Estimation for general number of external RWDs}
\usage{
rweWL_G(lst_dta, lambda, type = c("continuous", "binary"))
}
\arguments{
\item{lambda}{number of patients to be included or borrowed. For current
study, it should be equal to the number of patients}

\item{type}{type of outcomes}

\item{dta}{list of data from each source including current study}
}
\description{
Weighted Likelihood Estimation for general number of external RWDs
}
