% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweXBeta}
\alias{rweXBeta}
\title{Simulate X*Beta}
\usage{
rweXBeta(nPat, muCov, sdCov, corCov, regCoeff, cov.x = NULL, fmla = NULL)
}
\arguments{
\item{nPat}{number of patients}

\item{muCov}{mean vector of covariates}

\item{sdCov}{standard deviation vector of covariates}

\item{corCov}{correlation of covariates}

\item{regCoeff}{regression coefficients}
}
\description{
Simulate X*Beta
}
