% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_roche_palette.R
\name{show_roche_palette}
\alias{show_roche_palette}
\title{Show Company Categorical Colour Palette}
\usage{
show_roche_palette(title = "Company Categorical Palette", ncol = 7)
}
\arguments{
\item{title}{A string for the plot title. Default
`"Company Categorical Palette"`.}

\item{ncol}{Number of columns to arrange swatches in. Default `7`.}
}
\value{
A `ggplot` object.
}
\description{
Displays the company's 14-colour categorical palette as a labelled tile
plot, showing both the category name and hex code for each colour.
}
\examples{
\dontrun{
show_roche_palette()
show_roche_palette(ncol = 4)
}
}
