% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_package.R
\name{simupara}
\alias{simupara}
\title{Parameters for simulating data}
\arguments{
\item{nPat}{number of patients}

\item{muCov}{mean vector of covariates}

\item{sdCov}{standard deviation vector of covariates}

\item{corCov}{correlation of covariates}

\item{regCoeff}{regression coefficients}

\item{type}{distributions of the random error}

\item{ysig}{standard error of the random error}

\item{skew.n}{parameter of negative bionomial distribution}

\item{skew.p}{parameter of negative bionomial distribution}

\item{b0}{intercept in regession model}

\item{bin.mu}{mean of the binary outcomes used to compute b0}

\item{formula.z}{formula of the treatement assignment model. No intercept term.}

\item{formula.y}{formula of the outcome model. No intercept term.}
}
\description{
Parameters for simulating data
}
