% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_present.R
\name{summary2.RWE_DWITHPS}
\alias{summary2.RWE_DWITHPS}
\title{Summary unbalance metrics}
\usage{
\method{summary2}{RWE_DWITHPS}(x, v.cov = NULL, label.cov = v.cov, ...)
}
\description{
Summary unbalance metrics
}
