% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{tk_extract_stats}
\alias{tk_extract_stats}
\title{Extract summary statistics from a data frame}
\usage{
tk_extract_stats(x, quants = c(0.5), xlev = NULL, type = NULL, weights = NULL)
}
\arguments{
\item{quants}{quantiles for continuous variables}

\item{dta}{data frame}
}
\value{
a list of summary statistics
}
\description{
Extract summary statistics from a data frame
}
