% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{tkExtractStats}
\alias{tkExtractStats}
\title{Extract summary statistics from a data frame}
\usage{
tkExtractStats(x, quants = c(0.5), xlev = NULL, type = NULL)
}
\arguments{
\item{quants}{quantiles for continuous variables}

\item{dta}{data frame}
}
\value{
a list of summary statistics
}
\description{
Extract summary statistics from a data frame
}
