% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_stan.R
\name{rweDrawPost}
\alias{rweDrawPost}
\title{Get Posterior for all stratum}
\usage{
rweDrawPost(data, v.outcome = "Y", As = 0, type = c("continuous",
  "binary"), ...)
}
\arguments{
\item{data}{class DWITHPS data frame}

\item{As}{power of the power prior for each strata}

\item{...}{extra parameters for calling function \code{\link{rweSTAN}}}
}
\description{
Get Posterior for all stratum
}
