% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_wl.R
\name{rwePsWl}
\alias{rwePsWl}
\title{PS-Integrated Weighted Likelihood Estimation}
\usage{
rwePsWl(
  data,
  lambdas,
  v.outcome = "Y",
  m.var = c("jk", "bs"),
  bs.n = 1000,
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{class DWITHPS data frame}

\item{lambdas}{power parameter without standardization by ns0}

\item{m.var}{method to get variance: jackknife or bootstrap}

\item{bs.n}{number of bootstraps}

\item{seed}{random seed}

\item{...}{parameters for \code{rweWL}}
}
\description{
For all stratum. Variance estimated by bootstrap or Jack-Knife. Works for
one- or two-arm studies when there is one RWD
}
