% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweSimuFromTrial}
\alias{rweSimuFromTrial}
\title{Simulate data from an existing dataset}
\usage{
rweSimuFromTrial(trial.data, sizes, group = "A", outcome = "Y",
  keep.arm = TRUE, trt.effect = 0, with.replacement = FALSE,
  seed = NULL)
}
\description{
Simulate data from an existing dataset
}
