% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweSimuTwoArm}
\alias{rweSimuTwoArm}
\title{Simulate continuous outcomes for a two arm study}
\usage{
rweSimuTwoArm(nPat, muCov, sdCov, corCov, regCoeff.y, regCoeff.z = 0,
  fmla.y = NULL, fmla.z = NULL, ysig = NULL, b0 = NULL, sig2Ratio = 2,
  bin.mu = 0.5, ..., do.simu = TRUE)
}
\arguments{
\item{nPat}{number of patients}

\item{muCov}{mean vector of covariates}

\item{sdCov}{standard deviation vector of covariates}

\item{corCov}{correlation of covariates}

\item{ysig}{standard error of the random error}

\item{b0}{intercept in regession model}

\item{bin.mu}{mean of the binary outcomes used to compute b0}
}
\description{
Simulate continuous outcomes for a two arm study
}
